requirements.txt
# Core Framework
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings

# RAG & LangChain
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-core

# Vector Database
faiss-cpu

# PDF Processing
PyPDF2

# UI Framework
streamlit

# Utilities
python-dotenv
numpy

# HTTP Client
httpx
requests

# Logging
python-json-logger
